+++ resources/views/admin/media/index.blade.php
@@ -33,7 +33,3 @@
 
 <x-media.edit-modal />
 @endsection
-
-@push('scripts')
-<script defer src="https://cdn.jsdelivr.net/npm/alpinejs@3.x.x/dist/cdn.min.js"></script>
-@endpush