+++ resources/views/admin/plugins/index.blade.php
@@ -161,7 +161,3 @@
     }
 </style>
 @endpush
-@push('scripts')
-{{-- Alpine CDN --}}
-<script defer src="https://cdn.jsdelivr.net/npm/alpinejs@3.x.x/dist/cdn.min.js"></script>
-@endpush