+++ resources/views/components/media/edit-modal.blade.php
@@ -365,6 +365,4 @@ function mediaEdit(id, csrf) {
     }
 }
 </script>
-
-<script defer src="https://cdn.jsdelivr.net/npm/alpinejs@3.x.x/dist/cdn.min.js"></script>
 @endpush