+++ resources/views/components/media/grid.blade.php
@@ -426,6 +426,4 @@ function mediaGridComponent(config) {
     }
 }
 </script>
-
-<script defer src="https://cdn.jsdelivr.net/npm/alpinejs@3.x.x/dist/cdn.min.js"></script>
 @endpush